requirements.txt
Django>=5.1.0
djangorestframework>=3.15.0
django-environ>=0.10.0
httpx>=0.27.0
PyJWT>=2.9.0